Rename package properly
